chore: dissolve `converter`
