--- packages/shared/lib/chrysalis.json
@@ -1,3 +1,0 @@
-{
-    "snapshot": false
-}
