+++ packages/shared/lib/core/wallet/actions/mintNativeToken.ts
@@ -2,7 +2,7 @@ import { get } from 'svelte/store'
 import { selectedAccount, updateSelectedAccount } from '@core/account'
 import { localize } from '@core/i18n'
 import { NativeTokenOptions, TransactionOptions } from '@iota/wallet'
-import { Converter } from '@lib/converter'
+import { Converter } from '@core/utils'
 import { showAppNotification } from '@auxiliary/notification'
 import { activeProfile, ProfileType } from '@core/profile'
 import { handleLedgerError } from '@core/ledger'
