+++ packages/shared/lib/core/wallet/utils/convertHexAddressToBech32.ts
@@ -1,6 +1,6 @@
 import { networkHrp } from '@core/network'
 import { Bech32Helper } from '@core/utils'
-import { Converter } from '@lib/converter'
+import { Converter } from '@core/utils'
 import { get } from 'svelte/store'
 
 export function convertHexAddressToBech32(addressType: number, hexAddress: string): string {
