+++ packages/shared/lib/core/wallet/utils/outputs/getMetadataFromOutput.ts
@@ -1,6 +1,6 @@
 import { FEATURE_TYPE_METADATA, OUTPUT_TYPE_TREASURY } from '../../constants'
 import { IMetadataFeature, OutputTypes } from '@iota/types'
-import { Converter } from '@lib/converter'
+import { Converter } from '@core/utils'
 
 export function getMetadataFromOutput(output: OutputTypes): string {
     if (output && output?.type !== OUTPUT_TYPE_TREASURY) {
