+++ packages/shared/lib/core/wallet/utils/outputs/getNftActivityData.ts
@@ -12,7 +12,7 @@ import { ActivityDirection, ActivityType } from '@core/wallet/enums'
 import { ADDRESS_TYPE_NFT, EMPTY_HEX_ID, OUTPUT_TYPE_NFT } from '@core/wallet/constants'
 import { IAccountState } from '@core/account'
 import type { IMetadataFeature, INftOutput } from '@iota/types'
-import { Converter } from '@lib/converter'
+import { Converter } from '@core/utils'
 import { getAsyncDataFromOutput } from './getAsyncDataFromOutput'
 import { MimeType } from '@core/wallet/types'
 import { Blake2b } from '@iota/crypto.js'
