+++ packages/shared/lib/core/wallet/utils/outputs/getTagFromOutput.ts
@@ -1,6 +1,6 @@
 import { FEATURE_TYPE_TAG, OUTPUT_TYPE_TREASURY } from '../../constants'
 import { ITagFeature, OutputTypes } from '@iota/types'
-import { Converter } from '@lib/converter'
+import { Converter } from '@core/utils'
 
 export function getTagFromOutput(output: OutputTypes): string {
     if (output && output?.type !== OUTPUT_TYPE_TREASURY) {
